rename street to street_pattern
