+++ src/locales/ro_MD/location/index.ts
@@ -13,9 +13,9 @@ import postcode from './postcode';
 import secondary_address from './secondary_address';
 import state from './state';
 import state_abbr from './state_abbr';
-import street from './street';
 import street_address from './street_address';
 import street_name_part from './street_name_part';
+import street_pattern from './street_pattern';
 import street_prefix from './street_prefix';
 
 const location: LocationDefinitions = {
@@ -29,9 +29,9 @@ const location: LocationDefinitions = {
   secondary_address,
   state,
   state_abbr,
-  street,
   street_address,
   street_name_part,
+  street_pattern,
   street_prefix,
 };
