+++ src/locales/ro_MD/location/street_pattern.ts
